apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.13
pydantic-settings==2.10.1
passlib[argon2,bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
email-validator==2.2.0
httpx==0.27.0
openai
ollama
stripe==12.4.0
alembic==1.16.4
pyttsx3==2.90
pydub==0.25.1
mutagen==1.46.0
authlib==1.3.0
pytest==8.4.1
